Make exporting as file a feature of the calibration flavor
