+++ app/src/calibration/java/org/covidwatch/android/di/FlavorSpecificModules.kt
@@ -1,0 +1,16 @@
+package org.covidwatch.android.di
+
+import org.covidwatch.android.ui.reporting.NotifyOthersViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val flavorSpecificModule = module {
+    viewModel {
+        NotifyOthersViewModel(
+            startUploadDiagnosisKeysWorkUseCase = get(),
+            exportDiagnosisKeysAsFileUseCase = get(),
+            enManager = get(),
+            positiveDiagnosisRepository = get()
+        )
+    }
+}
