+++ app/src/calibration/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+    <!-- MENU -->
+    <string name="menu_reset_possible_exposures">[DEMO] Reset Possible Exposures</string>
+    <string name="menu_detect_exposures_from_server">[DEMO] Detect Exposures from Server</string>
+    <string name="menu_set_exposure_configuration">[DEMO] Set Exposure Configuration JSON</string>
+</resources>
