+++ app/src/main/java/org/covidwatch/android/di/AppModule.kt
@@ -33,7 +33,6 @@ import org.covidwatch.android.ui.exposures.ExposuresViewModel
 import org.covidwatch.android.ui.home.HomeViewModel
 import org.covidwatch.android.ui.menu.MenuViewModel
 import org.covidwatch.android.ui.onboarding.EnableExposureNotificationsViewModel
-import org.covidwatch.android.ui.reporting.NotifyOthersViewModel
 import org.covidwatch.android.ui.settings.SettingsViewModel
 import org.koin.android.ext.koin.androidApplication
 import org.koin.android.ext.koin.androidContext
@@ -238,15 +237,6 @@ val appModule = module {
 
     viewModel {
         MenuViewModel(exposureInformationRepository = get())
-    }
-
-    viewModel {
-        NotifyOthersViewModel(
-            startUploadDiagnosisKeysWorkUseCase = get(),
-            exportDiagnosisKeysAsFileUseCase = get(),
-            enManager = get(),
-            positiveDiagnosisRepository = get()
-        )
     }
 
     single {
