+++ app/src/production/java/org/covidwatch/android/di/FlavorSpecificModules.kt
@@ -1,0 +1,13 @@
+package org.covidwatch.android.di
+
+import org.koin.dsl.module
+
+val flavorSpecificModule = module {
+    viewModel {
+        NotifyOthersViewModel(
+            startUploadDiagnosisKeysWorkUseCase = get(),
+            enManager = get(),
+            positiveDiagnosisRepository = get()
+        )
+    }
+}
