+++ app/src/production/java/org/covidwatch/android/ui/reporting/NotifyOthers.kt
@@ -1,0 +1,18 @@
+package org.covidwatch.android.ui.reporting
+
+import org.covidwatch.android.data.positivediagnosis.PositiveDiagnosisRepository
+import org.covidwatch.android.domain.ExportDiagnosisKeysAsFileUseCase
+import org.covidwatch.android.domain.StartUploadDiagnosisKeysWorkUseCase
+import org.covidwatch.android.exposurenotification.ExposureNotificationManager
+
+class NotifyOthersFragment : BaseNotifyOthersFragment()
+
+class NotifyOthersViewModel(
+    startUploadDiagnosisKeysWorkUseCase: StartUploadDiagnosisKeysWorkUseCase,
+    enManager: ExposureNotificationManager,
+    positiveDiagnosisRepository: PositiveDiagnosisRepository
+) : BaseNotifyOthersViewModel(
+    startUploadDiagnosisKeysWorkUseCase,
+    enManager,
+    positiveDiagnosisRepository
+)
